ExprStoredEnchant - modernize
