+++ src/main/java/com/shanebeestudios/skbee/elements/other/OtherElementRegistration.java
@@ -19,6 +19,7 @@ import com.shanebeestudios.skbee.elements.other.effects.EffBlockLock;
 import com.shanebeestudios.skbee.elements.other.effects.EffBlockRandomlyTick;
 import com.shanebeestudios.skbee.elements.other.effects.EffBlockstateUpdate;
 import com.shanebeestudios.skbee.elements.other.effects.EffBreakBlocksWithEffects;
+import com.shanebeestudios.skbee.elements.other.effects.EffChunkLoad;
 import com.shanebeestudios.skbee.elements.other.effects.EffChunkRefresh;
 import com.shanebeestudios.skbee.elements.other.effects.EffChunkUnload;
 import com.shanebeestudios.skbee.elements.other.effects.EffDispatchCommand;
@@ -30,7 +31,6 @@ import com.shanebeestudios.skbee.elements.other.effects.EffEquipmentChange;
 import com.shanebeestudios.skbee.elements.other.effects.EffFreezePlayerConnection;
 import com.shanebeestudios.skbee.elements.other.effects.EffGiveOrDrop;
 import com.shanebeestudios.skbee.elements.other.effects.EffHurtAnimation;
-import com.shanebeestudios.skbee.elements.other.effects.EffChunkLoad;
 import com.shanebeestudios.skbee.elements.other.effects.EffOpenContainerAnimation;
 import com.shanebeestudios.skbee.elements.other.effects.EffOpenRealInventory;
 import com.shanebeestudios.skbee.elements.other.effects.EffOpenSign;
@@ -152,7 +152,6 @@ import com.shanebeestudios.skbee.elements.other.expressions.ExprSpawnerSpawnCoun
 import com.shanebeestudios.skbee.elements.other.expressions.ExprSpawnerSpawnDelay;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprSpawnerSpawnRange;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprSpellcasterSpell;
-import com.shanebeestudios.skbee.elements.other.expressions.ExprStoredEnchant;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprSuspiciousBlock;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprTabCompletion;
 import com.shanebeestudios.skbee.elements.other.expressions.ExprTabCompletionArgs;
@@ -175,10 +174,10 @@ import com.shanebeestudios.skbee.elements.other.type.BlockTypes;
 import com.shanebeestudios.skbee.elements.other.type.Comps;
 import com.shanebeestudios.skbee.elements.other.type.EntityTypes;
 import com.shanebeestudios.skbee.elements.other.type.EventTypes;
+import com.shanebeestudios.skbee.elements.other.type.Functions;
 import com.shanebeestudios.skbee.elements.other.type.ItemTypes;
 import com.shanebeestudios.skbee.elements.other.type.PlayerTypes;
 import com.shanebeestudios.skbee.elements.other.type.ServerTypes;
-import com.shanebeestudios.skbee.elements.other.type.Functions;
 import com.shanebeestudios.skbee.elements.other.type.WorldTypes;
 
 public class OtherElementRegistration {
@@ -343,7 +342,6 @@ public class OtherElementRegistration {
         ExprSpawnerSpawnDelay.register(registration);
         ExprSpawnerSpawnRange.register(registration);
         ExprSpellcasterSpell.register(registration);
-        ExprStoredEnchant.register(registration);
         ExprSuspiciousBlock.register(registration);
         ExprTabCompletion.register(registration);
         ExprTabCompletionArgs.register(registration);
